+++ src/dgbowl_schemas/yadg/dataschema_4_1/__init__.py
@@ -1,1 +1,9 @@
-from .schema import DataSchema
+from pydantic import BaseModel, Extra
+from typing import Sequence
+from .metadata import Metadata
+from .step import Steps
+
+
+class DataSchema(BaseModel, extra=Extra.forbid):
+    metadata: Metadata
+    steps: Sequence[Steps]
